requirements.txt
dify_plugin>=0.4.0,<0.7.0
kaggle
python-slugify
